llm-api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
python-dotenv==1.0.1
langchain==0.2.10
langchain-google-genai==2.0.0